Implementada cola, implementado bien controles de nextSong y previousSong, implementada bien la library de canciones favoritas del usuario y mejoras de UI implementadas.- By Daniel Lillo

+++ src/main/java/com/sonik/service/impl/LibraryServiceImpl.java
@@ -18,12 +18,19 @@ public class LibraryServiceImpl implements LibraryService {
 
     @Override
     public void addFavouriteSong(Song song) throws DuplicateIdException, DataAccessException, IncorrectArgumentException, ObjectNotFoundException {
-
         User user = AppContext.getUserService().getPreferences().getUser();
 
-        if (AppContext.getJpaUserLibraryRepository().exists(user.getId(), song.getId()))
-            return; //  Ya existe
+        //  Verificar si la canción ya existe en la tabla Song poe Url publica
+        if(!AppContext.getJpaSongRepository().existsUrl(song.getOriginalUrl())){
+            AppContext.getJpaSongRepository().save(song);
+        }
 
+        //  Verificar si ya es favorito
+        if (isFavourite(song)) {
+            return;
+        }
+
+        //  Crear relación en UserLibrary
         UserLibrary entry = new UserLibrary(user, song, LocalDate.now());
         AppContext.getJpaUserLibraryRepository().create(entry);
     }
@@ -31,8 +38,13 @@ public class LibraryServiceImpl implements LibraryService {
     @Override
     public void removeFavouriteSong(Song song) {
         try {
+            // Asegurar que tenemos la canción persistida
+            Song persistedSong = AppContext.getJpaSongRepository().findByUrl(song.getOriginalUrl());
+            if (persistedSong == null) {
+                return;
+            }
             User user = AppContext.getUserService().getPreferences().getUser();
-            UserLibraryId id = new UserLibraryId(user.getId(), song.getId());
+            UserLibraryId id = new UserLibraryId(user.getId(), persistedSong.getId());
             AppContext.getJpaUserLibraryRepository().delete(id);
         } catch (Exception ignored) {}
     }
@@ -50,4 +62,17 @@ public class LibraryServiceImpl implements LibraryService {
             return List.of();
         }
     }
+
+    @Override
+    public boolean isFavourite(Song song) throws DataAccessException, ObjectNotFoundException {
+        // Buscar la canción persistida por URL
+        Song persistedSong = AppContext.getJpaSongRepository().findByUrl(song.getOriginalUrl());
+        if (persistedSong == null) {
+            return false;
+        }
+
+        int userId = AppContext.getUserService().getPreferences().getUser().getId();
+        return AppContext.getJpaUserLibraryRepository().exists(userId, persistedSong.getId());
+    }
+
 }
